Added basic goalie summary parsing
Initial commit towards game summary parsing - adding in goalies. I'm not
sure what the end goal of this library might be, so I will likely move
onto other items before continuing the parsing of this page.

Much of it might be (or certainly is) duplicated on other reports.
Should overlapping data in Event Summary and Game Summary be present in
both game.event_summary and game.game_summary? Should we merge the two
and move to just game.summary? Plenty of options and I'll wait for that
before moving too far forward.